+++ app/services/weeks/finish_service.rb
@@ -2,36 +2,15 @@
 
 module Weeks
   class FinishService
-    prepend ApplicationService
-
-    def initialize(
-      bench_substitutions_service: Lineups::BenchSubstitutionsService
-    )
-      @bench_substitutions_service = bench_substitutions_service
-    end
+    include Deps[bench_substitutions_service: 'services.weeks.bench_substitutions']
 
     def call(week:)
       return if week.nil?
       return unless week.active?
 
+      bench_substitutions_service.call(week: week)
       # commento: weeks.status
       week.update!(status: Week::FINISHED)
-      make_bench_substitutions(week)
-    end
-
-    private
-
-    def make_bench_substitutions(week)
-      return unless Sport.find_by(title: week.league.sport_kind).changes
-
-      week.lineups.each do |lineup|
-        @bench_substitutions_service.call(lineup: lineup)
-      end
-
-      ::Lineups::Players::Points::UpdateJob.perform_later(
-        team_player_ids: week.teams_players.ids,
-        week_id: week.id
-      )
     end
   end
 end